dev-requirements.txt
numpy
cython
nose
wheel
pillow